Get absolute path for preload lib and added link check

+++ scripts/mysqld_safe.sh
@@ -278,6 +278,7 @@ parse_arguments() {
 # LD_LIBRARY_PATH and stripped from the lib value.
 add_mysqld_ld_preload() {
   lib_to_add="$1"
+  lib_to_add=$(readlink -f $lib_to_add)
   log_notice "Adding '$lib_to_add' to LD_PRELOAD for mysqld"
 
   # Check if the library is in the reduced number of standard system directories
@@ -666,9 +667,9 @@ then
     logging=file
   fi
 
-  if [ ! -f "$err_log" ]; then                  # if error log already exists,
-    touch "$err_log"                            # we just append. otherwise,
-    chmod "$fmode" "$err_log"                   # fix the permissions here!
+  if [ ! -f "$err_log" -a ! -h "$err_log" ]; then  # if error log already exists,
+    touch "$err_log"                               # we just append. otherwise,
+    chmod "$fmode" "$err_log"                      # fix the permissions here!
   fi
 
 fi
@@ -681,7 +682,7 @@ then
     USER_OPTION="--user=$user"
   fi
   # Change the err log to the right user, if it is in use
-  if [ $want_syslog -eq 0 ]; then
+  if [ $want_syslog -eq 0 -a ! -h "$err_log" ]; then
     touch "$err_log"
     chown $user "$err_log"
   fi
@@ -701,9 +702,11 @@ safe_mysql_unix_port=${mysql_unix_port:-${MYSQL_UNIX_PORT:-@MYSQL_UNIX_ADDR@}}
 mysql_unix_port_dir=`dirname $safe_mysql_unix_port`
 if [ ! -d $mysql_unix_port_dir ]
 then
-  mkdir $mysql_unix_port_dir
-  chown $user $mysql_unix_port_dir
-  chmod 755 $mysql_unix_port_dir
+  if [ ! -h $mysql_unix_port_dir ]; then
+    mkdir $mysql_unix_port_dir
+    chown $user $mysql_unix_port_dir
+    chmod 755 $mysql_unix_port_dir
+  fi
 fi
 
 # If the user doesn't specify a binary, we assume name "mysqld"
@@ -822,7 +825,11 @@ then
       exit 1
     fi
   fi
-  rm -f "$pid_file"
+
+  if [ ! -h "$pid_file" ]; then
+    rm -f "$pid_file"
+  fi
+
   if test -f "$pid_file"
   then
     log_error "Fatal error: Can't remove the pid file:
@@ -957,12 +964,20 @@ have_sleep=1
 while true
 do
   # Some extra safety
-  rm -f $safe_mysql_unix_port "$pid_file" "$pid_file.shutdown"	
+  if [ ! -h "$safe_mysql_unix_port" ]; then
+    rm -f "$safe_mysql_unix_port"
+  fi
+  if [ ! -h "$pid_file" ]; then
+    rm -f "$pid_file"
+  fi
+  if [ ! -h "$pid_file.shutdown" ]; then
+    rm -f "$pid_file.shutdown"
+  fi
   start_time=`date +%M%S`
 
   eval_log_error "$cmd"
 
-  if [ $want_syslog -eq 0 -a ! -f "$err_log" ]; then
+  if [ $want_syslog -eq 0 -a ! -f "$err_log" -a ! -h "$err_log" ]; then
     touch "$err_log"                    # hypothetical: log was renamed but not
     chown $user "$err_log"              # flushed yet. we'd recreate it with
     chmod "$fmode" "$err_log"           # wrong owner next time we log, so set
